+++ backend/src/main/java/com/summerHack/diningTogether/service/FoodService.java
@@ -8,17 +8,17 @@ import java.util.List;
 public abstract class FoodService {
     public abstract List<FoodBrief> getAllFood();
 
-    public abstract Food getFoodById(int id);
+    public abstract Food getFoodById(long id);
 
     public abstract String saveFood(Food food);
 
-    public abstract Food updateFood(int id, Food food);
+    public abstract Food updateFood(long id, Food food);
 
-    public abstract Food deleteById(Integer id);
+    public abstract Food deleteById(long id);
 
     public abstract List<FoodBrief> findByCategory(String category);
 
-    public abstract Food confirmFood(Integer id);
+    public abstract Food confirmFood(long id);
 
     public abstract List<FoodBrief> findAll();
 }
