+++ backend/src/main/java/com/summerHack/diningTogether/service/UserService.java
@@ -4,8 +4,8 @@ import com.summerHack.diningTogether.DTO.UserDTO;
 import com.summerHack.diningTogether.model.User;
 
 public abstract class UserService {
-    public abstract UserDTO getProfile(int id);
+    public abstract UserDTO getProfile(long id);
 
-    public abstract User update(int id, User user);
+    public abstract User update(long id, User user);
 
 }
